+++ src/components/common/hooks/useIncrementalResponse.ts
